% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogStuff.R
\name{blog_push}
\alias{blog_push}
\title{Push a blog post live (possibly)}
\usage{
blog_push(command = "git push")
}
\arguments{
\item{command}{Something which will be executed by \code{\link{system}}}
}
\value{
Used for its side effects.
}
\description{
I use this function to push blog posts live. This is an incredibyly lazy
wrapper for \code{blog_gen();system(command)}, where \code{command} is by
default how I upload changes to my website.
}

